feature add DI using Dagger2 and refactor Use cases
