+++ app/src/main/java/com/puzzlebench/clean_marvel_kotlin/presentation/MainActivity.kt
@@ -1,44 +1,26 @@
 package com.puzzlebench.clean_marvel_kotlin.presentation
 
 import android.os.Bundle
+import android.support.v7.app.AppCompatActivity
 import com.puzzlebench.clean_marvel_kotlin.R
-import com.puzzlebench.cmk.data.mapper.repository.CharacterMapperRepository
-import com.puzzlebench.cmk.data.repository.source.CharacterDataSourceImpl
-import com.puzzlebench.cmk.data.service.CharacterServicesImpl
-import com.puzzlebench.cmk.domain.usecase.GetCharacterRepositoryUseCase
-import com.puzzlebench.cmk.domain.usecase.GetCharacterServiceUseCase
-import com.puzzlebench.clean_marvel_kotlin.presentation.base.BaseRxActivity
 import com.puzzlebench.clean_marvel_kotlin.presentation.mvp.CharacterContract
-import com.puzzlebench.clean_marvel_kotlin.presentation.mvp.CharacterPresenter
-import com.puzzlebench.clean_marvel_kotlin.presentation.mvp.CharacterView
-import com.puzzlebench.cmk.data.repository.CharacterRepositoryImpl
-import com.puzzlebench.cmk.data.service.MarvelResquestGenerator
-import com.puzzlebench.cmk.domain.usecase.SaveCharacterRepositoryUseCase
 import dagger.android.AndroidInjection
 import javax.inject.Inject
 
-class MainActivity : BaseRxActivity() {
-
-    private val getCharacterServiceUseCase = GetCharacterServiceUseCase(CharacterServicesImpl(MarvelResquestGenerator.makeMarvelService()))
-    private val getCharacterRepositoryUseCase = GetCharacterRepositoryUseCase(CharacterRepositoryImpl(CharacterDataSourceImpl(), CharacterMapperRepository()))
-    private val saveCharacterRepositoryUseCase = SaveCharacterRepositoryUseCase(CharacterRepositoryImpl(CharacterDataSourceImpl(), CharacterMapperRepository()))
-
-    private val presenter = CharacterPresenter(CharacterView(this),
-            getCharacterServiceUseCase,
-            getCharacterRepositoryUseCase,
-            saveCharacterRepositoryUseCase,
-            subscriptions)
-
+class MainActivity : AppCompatActivity() {
 
     @Inject
-    lateinit var presenterV2: CharacterContract.Presenter
-
+    lateinit var presenter: CharacterContract.Presenter
 
     override fun onCreate(savedInstanceState: Bundle?) {
         AndroidInjection.inject(this)
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
-        //presenter.init()
-        presenterV2.initPresenter()
+        presenter.initPresenter()
+    }
+
+    override fun onPause() {
+        super.onPause()
+        presenter.onPause()
     }
 }
