--- app/src/main/java/com/puzzlebench/clean_marvel_kotlin/presentation/base/BasePresenter.kt
@@ -1,4 +1,0 @@
-package com.puzzlebench.clean_marvel_kotlin.presentation.base
-
-@Deprecated("")
-abstract class Presenter<out V>(val view: V)
