+++ app/src/main/java/com/puzzlebench/clean_marvel_kotlin/presentation/di/ActivityInjectorModule.kt
@@ -1,6 +1,7 @@
 package com.puzzlebench.clean_marvel_kotlin.presentation.di
 
 import com.puzzlebench.clean_marvel_kotlin.presentation.MainActivity
+import com.puzzlebench.clean_marvel_kotlin.presentation.di.android.ActivityScope
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
