+++ app/src/main/java/com/puzzlebench/clean_marvel_kotlin/presentation/di/CharacterModule.kt
@@ -2,8 +2,8 @@ package com.puzzlebench.clean_marvel_kotlin.presentation.di
 
 import com.puzzlebench.clean_marvel_kotlin.presentation.mvp.CharacterContract
 import com.puzzlebench.clean_marvel_kotlin.presentation.mvp.CharacterModel
-import com.puzzlebench.clean_marvel_kotlin.presentation.mvp.CharacterPresenterV2
-import com.puzzlebench.clean_marvel_kotlin.presentation.mvp.CharacterView2
+import com.puzzlebench.clean_marvel_kotlin.presentation.mvp.CharacterPresenter
+import com.puzzlebench.clean_marvel_kotlin.presentation.mvp.CharacterView
 import com.puzzlebench.cmk.data.di.repository.RepositoryModule
 import com.puzzlebench.cmk.data.di.service.ServiceModule
 import dagger.Binds
@@ -15,8 +15,8 @@ abstract class CharacterModule {
     abstract fun provideCharacterModel(characterModel: CharacterModel): CharacterContract.Model
 
     @Binds
-    abstract fun provideCharacterView2(characterView2: CharacterView2): CharacterContract.View
+    abstract fun provideCharacterView2(characterView2: CharacterView): CharacterContract.View
 
     @Binds
-    abstract fun provideCharacterPresenterV2(characterPresenterV2: CharacterPresenterV2): CharacterContract.Presenter
+    abstract fun provideCharacterPresenterV2(characterPresenter: CharacterPresenter): CharacterContract.Presenter
 }
