+++ app/src/main/java/com/puzzlebench/clean_marvel_kotlin/presentation/di/android/UiScopes.kt
@@ -1,4 +1,4 @@
-package com.puzzlebench.clean_marvel_kotlin.presentation.di
+package com.puzzlebench.clean_marvel_kotlin.presentation.di.android
 
 import javax.inject.Scope
