+++ app/src/main/java/com/puzzlebench/clean_marvel_kotlin/presentation/mvp/CharacterPresenter.kt
@@ -1,50 +1,37 @@
 package com.puzzlebench.clean_marvel_kotlin.presentation.mvp
 
-import com.puzzlebench.clean_marvel_kotlin.presentation.base.Presenter
-import com.puzzlebench.cmk.domain.model.Character
-import com.puzzlebench.cmk.domain.usecase.GetCharacterRepositoryUseCase
-import com.puzzlebench.cmk.domain.usecase.GetCharacterServiceUseCase
-import com.puzzlebench.cmk.domain.usecase.SaveCharacterRepositoryUseCase
-import io.reactivex.android.schedulers.AndroidSchedulers
-import io.reactivex.disposables.CompositeDisposable
-import io.reactivex.schedulers.Schedulers
+import io.reactivex.disposables.Disposable
+import javax.inject.Inject
 
-class CharacterPresenter constructor(view: CharacterView,
-                         private val getCharacterServiceUseCase: GetCharacterServiceUseCase,
-                         private val getCharacterRepositoryUseCase: GetCharacterRepositoryUseCase,
-                         private val saveCharacterRepositoryUseCase: SaveCharacterRepositoryUseCase,
-                         val subscriptions: CompositeDisposable) : Presenter<CharacterView>(view) {
+class CharacterPresenter @Inject constructor(private val view: CharacterContract.View,
+                                             private val model: CharacterContract.Model)
+    : CharacterContract.Presenter {
 
-    lateinit var characters: List<Character>
-    fun init() {
-        view.init()
-        characters = getCharacterRepositoryUseCase.invoke()
-        if (characters.isEmpty()) {
-            requestGetCharacters()
-        } else {
-            view.hideLoading()
-            view.showCharacters(characters)
-        }
+    private var subscription: Disposable? = null
+
+    override fun initPresenter() {
+        view.initView()
+        fetchData()
     }
 
-    private fun requestGetCharacters() {
-        val subscription = getCharacterServiceUseCase
-                .invoke()
-                .subscribeOn(Schedulers.io())
-                .observeOn(AndroidSchedulers.mainThread())
+    override fun fetchData() {
+        subscription = model
+                .fetchData()
                 .subscribe({ characters ->
-            if (characters.isEmpty()) {
-                view.showToastNoItemToShow()
-            } else {
-                saveCharacterRepositoryUseCase.invoke(characters)
-                view.showCharacters(characters)
-            }
-            view.hideLoading()
+                    if (characters.isEmpty()) {
+                        view.showMessageNoItemToShow()
+                    } else {
+                        view.showCharacters(characters)
+                    }
+                    view.hideLoading()
 
-        }, { e ->
-            view.hideLoading()
-            view.showToastNetworkError(e.message.toString())
-        })
-        subscriptions.add(subscription)
+                }, { e ->
+                    view.hideLoading()
+                    view.showNetworkError(e.message.toString())
+                })
+    }
+
+    override fun onPause() {
+        subscription?.dispose()
     }
 }
