+++ app/src/main/java/com/puzzlebench/clean_marvel_kotlin/presentation/mvp/CharacterView.kt
@@ -3,51 +3,51 @@ package com.puzzlebench.clean_marvel_kotlin.presentation.mvp
 import android.support.v7.widget.GridLayoutManager
 import android.view.View
 import com.puzzlebench.clean_marvel_kotlin.R
-import com.puzzlebench.cmk.domain.model.Character
 import com.puzzlebench.clean_marvel_kotlin.presentation.MainActivity
 import com.puzzlebench.clean_marvel_kotlin.presentation.adapter.CharacterAdapter
 import com.puzzlebench.clean_marvel_kotlin.presentation.extension.showToast
-import kotlinx.android.synthetic.main.activity_main.*
+import com.puzzlebench.cmk.domain.model.Character
+import kotlinx.android.synthetic.main.activity_main.progressBar
+import kotlinx.android.synthetic.main.activity_main.recycleView
 import java.lang.ref.WeakReference
+import javax.inject.Inject
 
-class CharacterView(activity: MainActivity) {
+class CharacterView @Inject constructor(private val activity: MainActivity) : CharacterContract.View {
+
+
     private val activityRef = WeakReference(activity)
     private val SPAN_COUNT = 1
-    var adapter = CharacterAdapter { character -> activity.applicationContext.showToast(character.name) }
+    private var adapter = CharacterAdapter { character -> activity.applicationContext.showToast(character.name) }
 
-    fun init() {
+    override fun initView() {
         val activity = activityRef.get()
-        if (activity != null) {
-            activity.recycleView.layoutManager = GridLayoutManager(activity, SPAN_COUNT)
-            activity.recycleView.adapter = adapter
+        activity?.let {
+            it.recycleView.layoutManager = GridLayoutManager(it, SPAN_COUNT)
+            it.recycleView.adapter = adapter
             showLoading()
         }
-
     }
 
-    fun showToastNoItemToShow() {
-        val activity = activityRef.get()
-        if (activity != null) {
-            val message = activity.baseContext.resources.getString(R.string.message_no_items_to_show)
-            activity.applicationContext.showToast(message)
-
-        }
+    override fun showLoading() {
+        activityRef.get()!!.progressBar.visibility = View.VISIBLE
     }
 
-    fun showToastNetworkError(error: String) {
-        activityRef.get()!!.applicationContext.showToast(error)
-    }
-
-    fun hideLoading() {
+    override fun hideLoading() {
         activityRef.get()!!.progressBar.visibility = View.GONE
     }
 
-    fun showCharacters(characters: List<Character>) {
-        adapter.data = characters
+    override fun showMessageNoItemToShow() {
+        activityRef.get()?.let {
+            val message = it.baseContext.resources.getString(R.string.message_no_items_to_show)
+            it.applicationContext.showToast(message)
+        }
     }
 
-    fun showLoading() {
-        activityRef.get()!!.progressBar.visibility = View.VISIBLE
+    override fun showNetworkError(error: String) {
+        activityRef.get()!!.applicationContext.showToast(error)
+    }
 
+    override fun showCharacters(characters: List<Character>) {
+        adapter.data = characters
     }
 }
