+++ data/src/main/java/com/puzzlebench/cmk/data/di/service/ApiModule.kt
@@ -1,6 +1,6 @@
 package com.puzzlebench.cmk.data.di.service
 
-import com.puzzlebench.cmk.data.service.MarvelResquestGenerator
+import com.puzzlebench.cmk.data.service.MarvelRequestGenerator
 import com.puzzlebench.cmk.data.service.api.MarvelApi
 import dagger.Module
 import dagger.Provides
@@ -9,7 +9,7 @@ import dagger.Provides
 class ApiModule {
     @Provides
     fun providesMarvelApi(): MarvelApi {
-        return MarvelResquestGenerator.makeMarvelService()
+        return MarvelRequestGenerator.makeMarvelService()
 
     }
 }
