+++ data/src/test/java/com/puzzlebench/cmk/data/mocks/factory/CharacterFactory.kt
@@ -12,7 +12,7 @@ class CharactersFactory {
         const val BASE_PATH = "image"
         const val BASE_EXTENSION = ".png"
 
-        fun getMockListCharacter(): List<Character> = listOf(1..5).map {
+        fun getMockListCharacter(): List<Character> = (1..5).map {
             Character("${BASE_NAME}$it", "${BASE_DESCRIPTION}$it", Thumbnail("${BASE_PATH}$it", BASE_EXTENSION))
         }
