--- domain/src/main/java/com/puzzlebench/cmk/domain/usecase/GetCharacterRepositoryUseCase.kt
@@ -1,11 +1,0 @@
-package com.puzzlebench.cmk.domain.usecase
-
-import com.puzzlebench.cmk.domain.model.Character
-import com.puzzlebench.cmk.domain.repository.CharacterRepository
-
-
-class GetCharacterRepositoryUseCase constructor(private val characterDataRepository: CharacterRepository) {
-
-    operator fun invoke(): List<Character> = characterDataRepository.getAll()
-
-}
